cur/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow>=2.6.0
torch>=1.9.0
scrapy>=2.5.0
beautifulsoup4>=4.9.3
requests>=2.26.0
flask>=2.0.1
python-dotenv>=0.19.0
pymongo>=3.12.0
scipy>=1.7.0
matplotlib>=3.4.3
seaborn>=0.11.2
joblib>=1.0.1 